refactors warranties_controller.py to display better results

+++ psycopg/controllers/warranties_controller.py
@@ -13,6 +13,17 @@ def warranty_create(request):
     product_id = post_data['product_id']
     warranty_months = post_data['warranty_months']
 
+    if not product_id or not warranty_months:
+        return jsonify({"message": "product_id and warranty_months are both required fields"}), 400
+
+    cursor.execute("""
+        SELECT * FROM Warranties
+        WHERE product_id=%s""",
+                   [product_id])
+    result = cursor.fetchone()
+    if result:
+        return jsonify({"message": f'warranty for product id {product_id} already exists'}), 400
+
     try:
         cursor.execute(
             """
@@ -24,9 +35,30 @@ def warranty_create(request):
         )
         conn.commit()
 
-        return jsonify({"message": f"Warranty for product id {product_id} has been added to the Warranties table."}), 200
+        cursor.execute(
+            """
+            SELECT *
+            FROM Warranties
+            ORDER BY warranty_id DESC;
+            """
+        )
+        result = cursor.fetchone()
 
-    except:
+        if result:
+            warranty_list = []
+
+            warranty_record = {
+                'warranty_id': result[0],
+                'product_id': result[1],
+                'warranty_months': result[2]
+            }
+
+            warranty_list.append(warranty_record)
+
+        return jsonify({"message": f"Warranty for product with id {product_id} has been added to the Warranties table.", "result": warranty_list}), 200
+
+    except Exception as e:
+        print(e)
         return jsonify({"message": "Warranty could not be added"}), 404
 
 
@@ -38,12 +70,25 @@ def warranties_get():
             FROM Warranties;
             """
         )
-        result = cursor.fetchall()
-        if result:
-            return jsonify(({"message": "warranties found", "result": result})), 200
+        results = cursor.fetchall()
+
+        if results:
+            warranty_list = []
+
+            for warranty in results:
+                warranty_record = {
+                    'warranty_id': warranty[0],
+                    'product_id': warranty[1],
+                    'warranty_months': warranty[2]
+                }
+
+                warranty_list.append(warranty_record)
+
+            return jsonify(({"message": "Warranties found", "results": warranty_list})), 200
         else:
             return jsonify(({"message": f"No warranties found"})), 404
-    except:
+    except Exception as e:
+        print(e)
         return jsonify({"message": "Could not fetch warranties data"}), 404
 
 
@@ -57,11 +102,23 @@ def warranty_get_by_id(warranty_id):
             """, [warranty_id]
         )
         result = cursor.fetchone()
+
         if result:
-            return jsonify(({"message": f"warranty with id {warranty_id} found", "result": result})), 200
+            warranty_list = []
+
+            warranty_record = {
+                'warranty_id': result[0],
+                'product_id': result[1],
+                'warranty_months': result[2]
+            }
+
+            warranty_list.append(warranty_record)
+
+            return jsonify(({"message": f"warranty with id {warranty_id} found", "result": warranty_list})), 200
         else:
             return jsonify(({"message": "No warranty found"})), 404
-    except:
+    except Exception as e:
+        print(e)
         return jsonify({"message": "Could not fetch warranty data"}), 404
 
 
@@ -80,10 +137,32 @@ def warranty_update_by_id(request, warranty_id):
                        [warranty_months, warranty_id])
         conn.commit()
 
-    except:
-        return jsonify({"message": "Product has been updated"}), 404
+        cursor.execute(
+            """
+            SELECT *
+            FROM Warranties
+            WHERE warranty_id=%s;
+            """, [warranty_id]
+        )
+        result = cursor.fetchone()
 
-    return jsonify({"message": f"{warranty_months} has been updated."}), 200
+        if result:
+            warranty_list = []
+
+            warranty_record = {
+                'warranty_id': result[0],
+                'product_id': result[1],
+                'warranty_months': result[2]
+            }
+
+            warranty_list.append(warranty_record)
+            return jsonify(({"message": f"warranty with id {warranty_id} has been updated", "result": warranty_list})), 200
+        else:
+            return jsonify(({"message": "No warranty found"})), 404
+
+    except Exception as e:
+        print(e)
+        return jsonify({"message": "Warranty could not be updated"}), 404
 
 
 def warranty_delete(warranty_id):
@@ -94,13 +173,26 @@ def warranty_delete(warranty_id):
         return jsonify({"message": 'warranty_id is required'}), 400
 
     try:
-        cursor.execute("""
-            DELETE FROM Warranties
-            WHERE warranty_id=%s
-            """, [warranty_id])
-        conn.commit()
+        cursor.execute(
+            """
+            SELECT *
+            FROM Warranties
+            WHERE warranty_id=%s;
+            """, [warranty_id]
+        )
+        result = cursor.fetchone()
 
-    except:
+        if result:
+            cursor.execute("""
+                DELETE FROM Warranties
+                WHERE warranty_id=%s
+                """, [warranty_id])
+            conn.commit()
+
+            return jsonify({"message": f"warranty with id {warranty_id} and its associated product have been deleted."}), 200
+        else:
+            return jsonify({"message": f"no warranty exists with id {warranty_id}"}), 404
+
+    except Exception as e:
+        print(e)
         return jsonify({"message": "warranty could not be deleted"}), 404
-
-    return jsonify({"message": f"warranty with id {warranty_id} has been deleted."}), 200
